+++ src/main/java/uniandes/edu/co/proyecto/Repositorio/ProductoRepository.java
@@ -1,6 +1,6 @@
 package uniandes.edu.co.proyecto.repositorio;
 
-import java.util.Date;
+import java.sql.Date;
 import java.util.Collection;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -31,7 +31,7 @@ public interface ProductoRepository extends JpaRepository <Producto, Integer>{
     @Modifying
     @Transactional
     @Query(value = "UPDATE productos SET nombre=:nombre, costo_bodega=:costo_bodega, precio_unitario=:precio_unitario, presentacion=:presentacion, peso=:peso,  volumen=:volumen, unidad_medida=:unidad_medida, cantidad_presentacion=:cantidad_presentacion,  fecha_vencimiento=:fecha_vencimiento, categoria=:categoria WHERE id = :id", nativeQuery = true)
-    void actualizarProducto(@Param("id") Long id, @Param("nombre") String nombre, @Param("costo_bodega") Integer costo_bodega, @Param("precio_unitario") Integer precio_unitario, 
+    void actualizarProducto(@Param("id") Integer id, @Param("nombre") String nombre, @Param("costo_bodega") Integer costo_bodega, @Param("precio_unitario") Integer precio_unitario, 
     @Param("presentacion") String presentacion, @Param("peso") Integer peso, @Param("volumen") Integer volumen, @Param("unidad_medida") String unidad_medida, 
     @Param("cantidad_presentacion") Integer cantidad_presentacion, @Param("fecha_vencimiento") Date fecha_vencimiento, @Param("categoria") Integer categoria);
